rename GarbageCollector to DeletedUsersIndex and move it to user subdir
